LayoutTests/platform/chromium-win/plugins/plugin-clip-subframe-expected.txt

CONSOLE MESSAGE: line 3: NPP_SetWindow: non-NULL window, Rect {0, 0, 300, 150}, Clip Rect {0, 0, 0, 0}, Type 1
CONSOLE MESSAGE: line 3: NPP_SetWindow: non-NULL window, Rect {0, 0, 300, 150}, Clip Rect {0, 0, 90, 150}, Type 1
